ClientStatistics + ExternalShares time column: integer -> bigint
The 'time' columns hold ms-since-epoch. SQLite's INTEGER is variable-
width and silently widens to 64-bit, so the upstream upstream-pool
declaration (type: 'integer') worked there. On Postgres 'integer' is
strictly int32 (max ~2.1 billion); ms-since-epoch overflowed in 1973.
Postgres rejected every share insert with
  value "1778639400000" is out of range for type integer (22003)
StratumV1ClientStatistics.addShares threw, the miner authorize cycle
saw repeated 500s, and no shares were credited.

Switch both 'time' columns to bigint. Other columns in the entity
tree (acceptedCount, shares as real, varchar lengths) are bounded
and stay as-is.

Note: the existing tables on the running Postgres still have the
int32 'time' columns from a previous synchronize: true round.
TypeORM's synchronize will not narrow/widen integer types on its
own; the deploy that picks up this commit must DROP TABLE
client_statistics_entity, external_shares_entity (and the indices
that reference 'time') so synchronize recreates them at the new
type. Stats history is fresh from this morning's deploy and not
worth preserving across this fix.

+++ src/ORM/client-statistics/client-statistics.entity.ts
@@ -22,7 +22,8 @@ export class ClientStatisticsEntity extends TrackedEntity {
     sessionId: string;
 
     @Index()
-    @Column({ type: 'integer' })
+    // ms-since-epoch — outgrows int32 in 1973-Apr; bigint required on Postgres.
+    @Column({ type: 'bigint' })
     time: number;
 
     @Column({ type: 'real' })
